+++ CHANGELOG.md
@@ -2,10 +2,29 @@
 
 ## Unreleased
 
-- The XDR definitions have been updated to support CAP-38 ([#451](https://github.com/stellar/js-stellar-base/pull/451)).
-- The `TrustLineAsset` and `ChangeTrustAsset` objects have been included to interact with both assets and liquidity pools ([#451](https://github.com/stellar/js-stellar-base/pull/451)).
-- Added helper `Operation.revokeLiquidityPoolSponsorship` ([#451](https://github.com/stellar/js-stellar-base/pull/451)).
-- Introduced new CAP-38 operations `LiquidityPoolDepositOp` and `LiquidityPoolWithdrawOp` ([#451](https://github.com/stellar/js-stellar-base/pull/451)).
+## [v6.0.0](https://github.com/stellar/js-stellar-base/compare/v5.3.2..v6.0.0)
+
+### Add
+
+- Introduced new CAP-38 operations `LiquidityPoolDepositOp` and `LiquidityPoolWithdrawOp`.
+- Introduced two new types of assets, `TrustLineAsset` and `ChangeTrustAsset`.
+
+### Update
+
+- The XDR definitions have been updated to support CAP-38.
+- Extended `Operation` class with the `Operation.revokeLiquidityPoolSponsorship` helper that allows revoking a liquidity pool sponsorship.
+- Asset types now include `AssetType.liquidityPoolShares`.
+
+### Breaking
+
+- The function `Operation.revokeTrustlineSponsorship` and the XDR objects `TrustLineEntry` and `LedgerKeyTrustLine` now use `TrustLineAsset` instead of `Asset`.
+- The function `Operation.changeTrust` and the XDR object `ChangeTrustOp` now use `ChangeTrustAsset` instead of `Asset`.
+- The `AssetAlphaNum4` and `AssetAlphaNum12` objects were renamed to `AlphaNum4` and `AlphaNum12` respectively.
+
+### Deprecated
+
+
+## [v5.3.2](https://github.com/stellar/js-stellar-base/compare/v5.3.1..v5.3.2)
 
 ### Fix
 - Update various dependencies to secure versions. Most are developer dependencies which means no or minimal downstream effects ([#446](https://github.com/stellar/js-stellar-base/pull/446), [#447](https://github.com/stellar/js-stellar-base/pull/447), [#392](https://github.com/stellar/js-stellar-base/pull/392), [#428](https://github.com/stellar/js-stellar-base/pull/428)); the only non-developer dependency upgrade is a patch version bump to `lodash` ([#449](https://github.com/stellar/js-stellar-base/pull/449)).